Moved DebugHelper (#1)
Moved DebugHelper from main index.js to Components and removed unnecessary div from index.jade
