+++ src/index.jade
@@ -27,8 +27,6 @@ html.no-js
                 a(href=LE_PRODUCTION_INSTANCE) siirry tuotantoon
         div
             #content
-        div
-            #debughelper
         if APP_MODE == 'development'
             script(src="/scripts/main.js", type="text/javascript")
